Add files via upload
This is a simple game that involves you guessing a number between 1 and 100, the terminal will tell you whether you are higher or lower than the number by range (for example your guess was lower by a range of 1 to 10). It is mainly an exercise in control flow with minor additions such as try and except.