Add files via upload
This is a simple version of the game hangman. You have 10 guesses and the word is randomly selected from a list of cities. It includes a countdown that ends the game if it hits zero, as well as a basic interface so you can see what letters you have got correct. It utilizes for loops and control flow as well as list and string editing. This was a lot of fun to put together!